+++ Walkie-iOS/Walkie-iOS/Sources/Extension/View+.swift
@@ -13,15 +13,6 @@ extension View {
         clipShape(RoundedCorner(radius: radius, corners: corners))
     }
     
-    func font(_ font: FontLevel) -> some View {
-        let fontSpacing = (font.lineHeight - font.fontSize) / 2
-        
-        return self
-            .font(Font.walkieFont(font))
-            .padding(.vertical, fontSpacing)
-            .lineSpacing(fontSpacing * 2)
-    }
-    
     func alignTo(_ alignment: Alignment) -> some View {
         modifier(AlignmentModifier(alignment: alignment))
     }
